+++ CHANGES.rst
@@ -11,6 +11,15 @@ https://github.com/zopefoundation/Zope/blob/4.0a6/CHANGES.rst
 4.0b5 (unreleased)
 ------------------
 
+New features
+++++++++++++
+
+- The `ProductContext` handed to a product's `initialize()` method
+  now has a `getApplication()` method which a product can use to,
+  e.g., add an object to the Application during startup (as used
+  by `Products.Sessions`).
+  (`#277 <https://github.com/zopefoundation/Zope/pull/277>`_)
+
 Bugfixes
 ++++++++